fixed byte issue

+++ src/test/java/inputs/test035/Test035.java
@@ -1,19 +1,14 @@
 package inputs.test035;
 
 class A {
-    int m(int b) {
-        System.out.println("A.m(byte)");
-        return b;
-    }
-    void m(double d) {
-        System.out.println("A.m(double)");
-    }
+    int m(byte b) { System.out.println("A.m(byte)"); return b; }
+    void m(double d) { System.out.println("A.m(double)"); }
 }
 
 public class Test035 {
     public static void main(String[] args) {
         A a = new A();
-        int b = 1;
+        byte b = 1;
         double d = 1.0;
         a.m(b);
         a.m(b + b);
